Technical interview preparation:
You are not allowed to google anything
Whiteboard first
Create a function def roman_to_int(roman_string): that converts a Roman numeral to an integer.

You can assume the number will be between 1 to 3999.

def roman_to_int(roman_string) must return an integer

If the roman_string is not a string or None, return 0